Add htaccess file to disallow any request to php file except index|matomo.php
In case a user is using apache and disallows any direct access to PHP files then this should avoid users having to whitelist these files.

Not sure if .htaccess files will actually be extracted on WordPress when they are present in a zip. In Matomo they aren't. We'll need to test this later I suppose to make Matomo work on more instances by default. Otherwise we should be able to listen to a WP hook and add it to the .htaccess file in the root directory.